backend/requirements.txt
# Workspace version: 1.42.61
ipykernel==6.29.5
fastapi==0.111.0
databutton==0.38.34
uvicorn[standard]==0.29.0
python-multipart==0.0.9
openai
beautifulsoup4
requests
PyExecJS